Courses from Turing Commons
Responsible Research and Innovation (RRI): will explore what it means to take (individual and collective) responsibility for (and over) the processes and outcomes of research and innovation in data science and AI.
Public Engagement of Data Science and AI (PED): designed to help you understand the practical and ethical value of public engagement with data science and AI.
AI Ethics & Governance (AEG): this course has practical and sometimes hands-on activities that are designed to a) encourage critical reflection and b) help you build practical understanding of the processes associated with effective and responsible engagement with AI systems. While they can be carried out as part of individual and self-directed learning, they are most suited to group discussion.
Data Justice (DJ): this course explores the emerging movement of data justice, which seeks to apply a social justice-oriented approach to examining the range of social, political and material concerns arising within our increasingly datafied society.